requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
line-bot-sdk==3.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask==3.0.0 